backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic[email]==2.6.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.27
psycopg[binary]==3.1.18
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
celery==5.3.6
redis==5.0.1
httpx==0.27.0
python-multipart==0.0.9
pyvips==2.2.2
Pillow==10.2.0
pytest==8.0.1
